Extract policy and profile CLI
